LayoutTests/imported/w3c/web-platform-tests/html/interaction/focus/focusgroup/tentative/backward-navigation/moves-to-previous-item-expected.txt

item1 item2

FAIL When the focus is set on a focusgroup item, an arrow key press should move the focus to the previous item. assert_equals: expected Element node <span id="item1" tabindex="0">item1</span> but got Element node <span id="item2" tabindex="-1">item2</span>
